Add a wrapper to the finally block of Start-DevPowershell which checks if ZapDisable is a null variable. This is to prevent an InvalidOperation error on exit.

+++ build.psm1
@@ -1971,8 +1971,10 @@ function Start-DevPowerShell {
             Remove-Item env:DEVPATH
         }
 
-        if ($ZapDisable) {
-            Remove-Item env:COMPLUS_ZapDisable
+        if (Get-Variable -Name ZapDisable -ValueOnly -ErrorAction SilentlyContinue) {
+            if ($ZapDisable) {
+                Remove-Item env:COMPLUS_ZapDisable
+            }
         }
     }
 }
